src/frontend/backend/requirements.txt
#For Automat-LLM
langchain_huggingface
langchain_community
langchain
transformers
sentence-transformers
pydantic
langchain_core
numpy
groq
rich
langchain_groq
dia
playsound

#Vector Storage (choose one)
faiss-cpu
# faiss-gpu

#For Automat-ImageGen
diffusers

#For automat-server
fastapi 
asyncio
uvicorn
uuid
rpds